Weight preprocessing
The header generation script now accepts weights taken directly from
keras. Previously, I had to modify the training script to get the
weights in the correct format, but it makes more sense to do it here
because the weight format is inherent to the implementation and not the
model itself.