Reduced LSTM firmware
This is tested to be working. In the future, the rank will be loaded
from the SD card as well. The idea is that all reduced models will use
the same firmware so we don't recompile every time we need to swap out
the model. We'll need to do this a lot. Maybe even 50 times.